Rework connection workflow.
Introduce a global state for VpnHandler class. This state is exported
and can be used outside of the class. We have a simple workflow:
disconnected, connecting, connected, disconnecting and disconnected.

A more detailed workflow can be overlaid on this workflow. This is not
used now but will be used when we will try to handle racoonctl
processes.

Use NSTimer to transition from one state to another to let the
interface a change to refresh appropriately.

Rework how we allow to close the application: the VPN should be
terminated.