Invoke racoonctl to make the connection.
racoonctl needs to be run as root because it needs to determine
properly the output IP and it is using an option which is not allowed
as a normal user.

The helper tool will open a PTY and give the password when it receives
the prompt. The helper tool will then pass back the file descriptor so
that we can interact with racoonctl. It also passes back a file
descriptor that will kill racoonctl if closed.

A timer is started when racoonctl is invoked. When elapsed, racoonctl
is killed and the connection is terminated. Data from racoonctl is
buffered in memory. When racoonctl closes its standard output, we look
at this buffer to check if the connection is successful. We get the IP
address.

Todo :
 - localization
 - bundle script
 - more racoonctl output parsing for more informative messages
 - phsae1 script should be able to handle DNS